+++ contracts/base/LowLevelErrorMessage.sol
@@ -5,11 +5,19 @@ abstract contract LowLevelErrorMessage {
     // __gap is reserved storage
     uint256[50] private __gap;
 
-    function _getRevertMessage(bytes memory reason) internal pure returns (string memory) {
-        if (reason.length < 68) return ("Unexpected error");
+    // suggested solution from ABDK, https://ethereum.stackexchange.com/a/110428/4955
+    function _getRevertMessage(bytes memory revertData) internal pure returns (string memory reason) {
+        uint256 len = revertData.length;
+        if (len < 68) return ("Unexpected error");
+        uint256 contentLength;
         assembly {
-            reason := add(reason, 0x04)
+            revertData := add(revertData, 4)
+            contentLength := mload(revertData) // Save the content of the length slot
+            mstore(revertData, sub(len, 4)) // Set proper length
         }
-        return (abi.decode(reason, (string)));
+        reason = abi.decode(revertData, (string));
+        assembly {
+            mstore(revertData, contentLength) // Restore the content of the length slot
+        }
     }
 }