+++ contracts/test/TestMetaTxRecipient.sol
@@ -2,13 +2,13 @@
 pragma solidity 0.7.6;
 
 import { Initializable } from "@openzeppelin/contracts-upgradeable/proxy/Initializable.sol";
-import { BaseRelayRecipient } from "../gsn/BaseRelayRecipient.sol";
+import { BaseRelayRecipient, IRelayRecipient } from "../gsn/BaseRelayRecipient.sol";
 
 contract TestMetaTxRecipient is BaseRelayRecipient, Initializable {
     address public pokedBy;
 
     function __TestMetaTxRecipient_init(address trustedForwarderArg) external initializer {
-        _trustedForwarder = trustedForwarderArg;
+        _setTrustedForwarder(trustedForwarderArg);
     }
 
     function poke() external {
@@ -16,7 +16,7 @@ contract TestMetaTxRecipient is BaseRelayRecipient, Initializable {
     }
 
     // solhint-disable
-    function error() external {
+    function error() external pure {
         revert("MetaTxRecipientMock: Error");
     }
 }