+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - rename `AccountBalance.getLiquidateMarginRequirement` to `AccountBalance.getMarginRequirementForLiquidation`
 
+### Removed
+- remove state `_versionRecipient` from `ClearingHouse` and `Vault`
 ## [0.12.6] -2021-10-25
 - bug fixes
   - rounding error at `ClearingHouse.closePosition()` and `Vault.withdraw()`